slack: add support for an external authentication script
When using the combined cookie + token auth mechanism, it is tedious to
poke around in browser state to find the token / cookie needed for a
slack account. This tedium is increased further when combined with short
life token/cookie expiry times.

This makes life more user friendly by introducing support for running an
external script to fetch tokens from the browser on the fly. As long as
the browser has a valid session, the slack plugin will "just work" with
no auth info required to be hardcoded.

The provided python script is copied from weechat's slack plugin with a
minor change to pass in the slack username and print out the token in a
space separated format.

The config option accepts a full command line string. Example simple
setting for the new 'auth_script' config option might

  /path/to/slack_purple_auth.py firefox

If pidgin is running inside a flatpak though it is possible to get
more advanced:

  flatpak-run --host /path/to/slack_purple_auth.py firefox

Or even more fancy if your primary env is within a toolbox container

  flatpak-run --host toolbox run /path/to/slack_purple_auth.py firefox

In addition to the lists set by the user in the settings, the account
username will be appended, so what actually gets run would be

  /path/to/slack_purple_auth.py firefox username%slack-workspace-hostname

All the script is really interested in is the 'slack-workspace-hostname'
part, which is matches against the 'url' field in the local config teams.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>